check if page is public before triggering redirect
